Refractored files to appropriate packages
