+++ app/src/main/java/com/metrostateics499/vre_app/view/CustomTextActivity.kt
@@ -1,9 +1,12 @@
-package com.metrostateics499.vre_app
+package com.metrostateics499.vre_app.view
 
 import android.os.Bundle
 import android.view.View
 import android.widget.*
 import androidx.appcompat.app.AppCompatActivity
+import com.metrostateics499.vre_app.model.data.CustomTextMessage
+import com.metrostateics499.vre_app.R
+import com.metrostateics499.vre_app.model.Passing
 
 class CustomTextActivity : AppCompatActivity(), CustomTextPopUps.Listener {
 
@@ -46,7 +49,7 @@ class CustomTextActivity : AppCompatActivity(), CustomTextPopUps.Listener {
             Toast.makeText(
                 this@CustomTextActivity,
                 "You have selected " +
-                    parent.getItemAtPosition(position),
+                        parent.getItemAtPosition(position),
                 Toast.LENGTH_SHORT
             ).show()
             textViewSelected = parent.getItemAtPosition(position).toString()
@@ -105,12 +108,16 @@ class CustomTextActivity : AppCompatActivity(), CustomTextPopUps.Listener {
             ).show()
             openPopUp(textViewSelected, "add")
         } else if (customTextString.isNotEmpty() &&
-            Passing.customTextMessageList.addCustomTextMessage(CustomTextMessage(customTextString))
+                Passing.customTextMessageList.addCustomTextMessage(
+                    CustomTextMessage(
+                        customTextString
+                    )
+                )
         ) {
             Toast.makeText(
                 this@CustomTextActivity,
                 "New Custom Text Message Successfully " +
-                    "Added",
+                        "Added",
                 Toast.LENGTH_SHORT
             ).show()
             refreshList()
@@ -139,7 +146,7 @@ class CustomTextActivity : AppCompatActivity(), CustomTextPopUps.Listener {
             ).show()
             openPopUp(textViewSelected, "edit")
         } else if (customTextString.isNotEmpty() &&
-            Passing.customTextMessageList.editCustomTextMessage(
+                Passing.customTextMessageList.editCustomTextMessage(
                     CustomTextMessage(textViewSelected),
                     customTextString
                 )
@@ -153,7 +160,7 @@ class CustomTextActivity : AppCompatActivity(), CustomTextPopUps.Listener {
             Toast.makeText(
                 this@CustomTextActivity,
                 "That custom text message already exists. " +
-                    "Try something else or click cancel.",
+                        "Try something else or click cancel.",
                 Toast.LENGTH_SHORT
             ).show()
             openPopUp(textViewSelected, "edit")
@@ -168,7 +175,7 @@ class CustomTextActivity : AppCompatActivity(), CustomTextPopUps.Listener {
             Toast.makeText(
                 this@CustomTextActivity,
                 "You have deleted custom text message: " +
-                    textViewSelected,
+                        textViewSelected,
                 Toast.LENGTH_SHORT
             ).show()
             refreshList()
