+++ app/src/main/java/com/metrostateics499/vre_app/view/KeyPhraseActivity.kt
@@ -7,7 +7,6 @@ import androidx.appcompat.app.AppCompatActivity
 import com.metrostateics499.vre_app.R
 import com.metrostateics499.vre_app.model.Passing
 import com.metrostateics499.vre_app.model.data.KeyPhrase
-import com.metrostateics499.vre_app.model.data.KeyPhrasePopUps
 
 /**
  * Key words activity
@@ -55,7 +54,7 @@ class KeyPhraseActivity : AppCompatActivity(), KeyPhrasePopUps.Listener {
             Toast.makeText(
                 this@KeyPhraseActivity,
                 "You have selected " +
-                    parent.getItemAtPosition(position),
+                        parent.getItemAtPosition(position),
                 Toast.LENGTH_SHORT
             ).show()
             textViewSelected = parent.getItemAtPosition(position).toString()
@@ -114,12 +113,12 @@ class KeyPhraseActivity : AppCompatActivity(), KeyPhrasePopUps.Listener {
             ).show()
             openPopUp(textViewSelected, "add")
         } else if (keyphraseString.isNotEmpty() &&
-            Passing.keyPhraseList.addKeyPhrase(KeyPhrase(keyphraseString))
+                Passing.keyPhraseList.addKeyPhrase(KeyPhrase(keyphraseString))
         ) {
             Toast.makeText(
                 this@KeyPhraseActivity,
                 "New Key Phrase Successfully " +
-                    "Added",
+                        "Added",
                 Toast.LENGTH_SHORT
             ).show()
             refreshList()
@@ -148,9 +147,9 @@ class KeyPhraseActivity : AppCompatActivity(), KeyPhrasePopUps.Listener {
             ).show()
             openPopUp(textViewSelected, "edit")
         } else if (keyphraseString.isNotEmpty() && Passing.keyPhraseList.editKeyPhrase(
-                KeyPhrase(textViewSelected),
-                keyphraseString
-            )
+                    KeyPhrase(textViewSelected),
+                    keyphraseString
+                )
         ) {
             Toast.makeText(
                 this@KeyPhraseActivity, "Successfully Edited",
@@ -161,7 +160,7 @@ class KeyPhraseActivity : AppCompatActivity(), KeyPhrasePopUps.Listener {
             Toast.makeText(
                 this@KeyPhraseActivity,
                 "That Key Phrase already exists. " +
-                    "Try something else or click cancel.",
+                        "Try something else or click cancel.",
                 Toast.LENGTH_SHORT
             ).show()
             openPopUp(textViewSelected, "edit")
@@ -174,7 +173,7 @@ class KeyPhraseActivity : AppCompatActivity(), KeyPhrasePopUps.Listener {
             Toast.makeText(
                 this@KeyPhraseActivity,
                 "You have deleted phrase: " +
-                    textViewSelected,
+                        textViewSelected,
                 Toast.LENGTH_SHORT
             ).show()
             refreshList()
